rename nodeIcons
