+++ pages/api/studentPortal/nodeIcons.ts
